backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
torch
torchvision
supabase
python-dotenv
numpy
moviepy
